+++ README.md
@@ -1,5 +1,5 @@
-# Pantry 🥘
+# `fagiolo` 🥘
 
 A weekly meal choice automator 🤖🥘
 
-Instead of having to try and remember which meals you like while planning your weekly menu, Pantry will choose for you from a selection of your pre-determined favourites
+Instead of having to try and remember which meals you like while planning your weekly menu, `fagiolo` will choose for you from a selection of your pre-determined favourites
