+++ fagiolo/app/__init__.py
@@ -11,4 +11,4 @@ def beautify_enum(enum_member: enum.Enum) -> str:
     return cast(str, enum_member.value.lower().replace("_", " "))
 
 
-from pantry.app import views  # noqa E402
+from fagiolo.app import views  # noqa E402
