+++ fagiolo/app/templates/base.html
@@ -11,7 +11,7 @@
       href="https://fonts.googleapis.com/css2?family=Roboto+Mono:wght@100;200&display=swap"
       rel="stylesheet"
     />
-    <title>Pantry</title>
+    <title>Fagiolo</title>
   </head>
   <body>
     <main class="content">{% block content %}{% endblock %}</main>
