+++ fagiolo/app/view_functions.py
