+++ fagiolo/fagiolo.py
@@ -1,0 +1,1 @@
+from fagiolo.app import app  # noqa
