--- pantry/pantry.py
@@ -1,1 +1,0 @@
-from pantry.app import app  # noqa
