+++ tests/conftest.py
@@ -1,15 +1,15 @@
 import pytest
 from flask.testing import FlaskClient
 
-from pantry import pantry
+from fagiolo import fagiolo
 
 
 @pytest.fixture
 def client() -> FlaskClient:
-    pantry.app.config["DEBUG"] = True
-    pantry.app.config["TESTING"] = True
-    pantry.app.config["PROPAGATE_EXCEPTIONS"] = False
-    pantry.app.config["PRESERVE_CONTEXT_ON_EXCEPTION"] = False
+    fagiolo.app.config["DEBUG"] = True
+    fagiolo.app.config["TESTING"] = True
+    fagiolo.app.config["PROPAGATE_EXCEPTIONS"] = False
+    fagiolo.app.config["PRESERVE_CONTEXT_ON_EXCEPTION"] = False
 
-    with pantry.app.test_client() as client:
+    with fagiolo.app.test_client() as client:
         yield client
